(refactor) Renamed GetOrderBook method to Get
