+++ exchanges/bitfinex/symbols.go
@@ -16,6 +16,17 @@ type SymbolsProvider struct {
 	httpClient *httpclient.Client
 }
 
+// Symbol - bitfinex symbol model
+type Symbol struct {
+	Pair           string `json:"pair"`
+	PricePrecision int64  `json:"price_precision"`
+	InitialMargin  string `json:"initial_margin"`
+	MinMargin      string `json:"minimum_margin"`
+	MaxOrderSize   string `json:"maximum_order_size"`
+	MinOrderSize   string `json:"minimum_order_size"`
+	Expiration     string `json:"expiration"`
+}
+
 // NewSymbolsProvider - SymbolsProvider constructor
 func NewSymbolsProvider(httpProxy proxy.Provider) *SymbolsProvider {
 	log.Println("Constructing symbols provider")
